refactor: Extract receive request entry types to dedicated header
Move QmlReceiveRequestRecipient and QmlRecentRequestEntry from an
anonymous namespace in walletqmlmodel.cpp into a dedicated header so
they can be shared with the upcoming ReceiveRequestHistoryModel.

Bump the recipient serialization version to 2, adding an optional
noteSelf field, and widen the date serialization type to int64_t to
avoid Y2038 truncation.
